% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football.R
\name{get_season}
\alias{get_season}
\title{Get Season Data}
\usage{
get_season(year, header)
}
\arguments{
\item{year}{An integer for season}

\item{header}{A header object}
}
\value{
A tibble with season related data along with attributes related to the competion and players links
}
\description{
Get Season Data
}
\examples{
TOKEN <- 'XXX'
s_2017 <- get_season(2017, header(TOKEN))

}
