% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football.R
\name{get_teams}
\alias{get_teams}
\title{Get teams on a competition}
\usage{
get_teams(competition_url, header)
}
\arguments{
\item{competition_url}{An url for the competion}

\item{header}{A header for the request}
}
\value{
A tibble with properties of competition along with attributes containing related links
}
\description{
Get teams on a competition
}
\examples{
TOKEN <- "XXX"
season_2017 <- get_season(2017, header(TOKEN))
teams_competition <- purrr::map_df(
    `names<-`(attr(season_2017, 'links')[['teams']], season_2017$id),
    ~get_teams(.x, header(TOKEN)),
    .id = "competition")
}
