% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football.R
\name{header}
\alias{header}
\title{Create Header}
\usage{
header(token = TOKEN)
}
\arguments{
\item{token}{A character string with api token created in openfootball}
}
\value{
A character vector with text required for header
}
\description{
Create Header
}
\examples{
\dontrun{
hdr <- header(token = "XXX")
}

}
