% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/football.R
\name{make_request}
\alias{make_request}
\title{Make a url request using the header}
\usage{
make_request(url, header)
}
\arguments{
\item{url}{A character of api url to make request}

\item{header}{The header to use to request the url}
}
\value{
A response object
}
\description{
Make a url request using the header
}
